refactor: organize node.js integration tests by service
Organizes the integration tests for Node.js by service using
service-specific folders. This allows us to run jest with a regex
pattern like "cache/" to match all cache tests; "topics/" to match all
topics-related tests, etc.

This greatly simplifies the CLI invocations to run service-specific
tests maintainably.
