+++ packages/client-sdk-nodejs/test/integration/shared/auth/auth-client.test.ts
@@ -1,4 +1,4 @@
-import {SetupAuthClientIntegrationTest} from '../integration-setup';
+import {SetupAuthClientIntegrationTest} from '../../integration-setup';
 import {runAuthClientTests} from '@gomomento/common-integration-tests';
 
 const {
