+++ packages/client-sdk-nodejs/test/integration/shared/cache/create-delete-list-cache.test.ts
@@ -1,5 +1,5 @@
 import {runCreateDeleteListCacheTests} from '@gomomento/common-integration-tests';
-import {SetupIntegrationTest} from '../integration-setup';
+import {SetupIntegrationTest} from '../../integration-setup';
 
 const {cacheClient} = SetupIntegrationTest();
