+++ packages/client-sdk-nodejs/test/integration/shared/cache/dictionary.test.ts
@@ -1,5 +1,5 @@
 import {runDictionaryTests} from '@gomomento/common-integration-tests';
-import {SetupIntegrationTest} from '../integration-setup';
+import {SetupIntegrationTest} from '../../integration-setup';
 
 const {cacheClient, integrationTestCacheName} = SetupIntegrationTest();
