+++ packages/client-sdk-nodejs/test/integration/shared/cache/item-get-ttl.test.ts
@@ -1,4 +1,4 @@
-import {SetupIntegrationTest} from '../integration-setup';
+import {SetupIntegrationTest} from '../../integration-setup';
 import {runItemGetTtlTest} from '@gomomento/common-integration-tests/dist/src/item-get-ttl';
 
 const {cacheClient, integrationTestCacheName} = SetupIntegrationTest();
