+++ packages/client-sdk-nodejs/test/integration/shared/cache/item-get-type.test.ts
@@ -1,4 +1,4 @@
-import {SetupIntegrationTest} from '../integration-setup';
+import {SetupIntegrationTest} from '../../integration-setup';
 import {runItemGetTypeTest} from '@gomomento/common-integration-tests';
 
 const {cacheClient, integrationTestCacheName} = SetupIntegrationTest();
