+++ packages/client-sdk-nodejs/test/integration/shared/cache/keys-exist.test.ts
@@ -1,5 +1,5 @@
 import {runKeysExistTest} from '@gomomento/common-integration-tests';
-import {SetupIntegrationTest} from '../integration-setup';
+import {SetupIntegrationTest} from '../../integration-setup';
 
 const {cacheClient, integrationTestCacheName} = SetupIntegrationTest();
