+++ packages/client-sdk-nodejs/test/integration/shared/leaderboard/leaderboard.test.ts
@@ -1,5 +1,5 @@
 import {runLeaderboardClientTests} from '@gomomento/common-integration-tests';
-import {SetupLeaderboardIntegrationTest} from '../integration-setup';
+import {SetupLeaderboardIntegrationTest} from '../../integration-setup';
 
 const {
   leaderboardClient,
