+++ packages/client-sdk-nodejs/test/integration/shared/topics/webhooks.test.ts
@@ -1,5 +1,5 @@
 import {runWebhookTests} from '@gomomento/common-integration-tests';
-import {SetupTopicIntegrationTest} from '../integration-setup';
+import {SetupTopicIntegrationTest} from '../../integration-setup';
 
 const {topicClient, cacheClient, integrationTestCacheName} =
   SetupTopicIntegrationTest();
